Refactor wsstream library from apiserver to apimachinery
