+++ staging/src/k8s.io/kubelet/pkg/cri/streaming/portforward/portforward.go
@@ -23,8 +23,8 @@ import (
 	"time"
 
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/httpstream/wsstream"
 	"k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apiserver/pkg/util/wsstream"
 )
 
 // PortForwarder knows how to forward content from a data stream to/from a port
