+++ staging/src/k8s.io/kubelet/pkg/cri/streaming/portforward/websocket.go
@@ -31,9 +31,9 @@ import (
 
 	api "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/httpstream/wsstream"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apiserver/pkg/endpoints/responsewriter"
-	"k8s.io/apiserver/pkg/util/wsstream"
 )
 
 const (
